requirements.txt
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.21.0
streamlit>=1.25.0
pandas>=1.5.0 